+++ Implicit-Lagrange/Topology_Optimization/Center_of_Mass_Constraint.h
@@ -109,7 +109,7 @@ private:
 
 public:
   bool nodal_density_flag_;
-  size_t last_comm_step, current_step, last_solve_step;
+  int last_comm_step, current_step, last_solve_step;
   std::string my_fea_module = "Inertial";
 
   CenterOfMassConstraint_TopOpt(FEA_Module *FEM, bool nodal_density_flag, int constraint_component, bool inequality_flag=true, real_t constraint_value=0)
