+++ Implicit-Lagrange/Topology_Optimization/Heat_Capacity_Potential_Minimize.h
@@ -106,7 +106,7 @@ private:
 
 public:
   bool nodal_density_flag_;
-  size_t last_comm_step, last_solve_step, current_step;
+  int last_comm_step, last_solve_step, current_step;
   std::string my_fea_module = "Heat_Conduction";
 
   HeatCapacityPotentialMinimize_TopOpt(FEA_Module *FEM, bool nodal_density_flag)
